examples/cross_region_adaptation/requirements.txt
xgboost>=2.0
scikit-learn>=1.4
pandas>=2.1
numpy>=1.26
pyyaml>=6.0
joblib>=1.3
shap>=0.44
optuna>=3.5
scipy>=1.12
